feat(payment): add customerCreated function
This commit adds the `customerCreated` function to handle the creation of a user when a new customer is created. It takes a JSON message as input, unmarshals it into a `stripe.Customer` struct, and then calls the `RegisterUser` function to create a new user in the system. The created user is then returned or an error is returned if any steps fail.