merge dev into main (#2)
* temp

* postman collection

* keep some files

* second commit

* Remove unnecessary job dependency from CI workflow

The commit removes the unnecessary job dependency from the CI workflow, improving the workflow's efficiency.

* Update ProxyConfig in main.go and rewrite ProxyHandler function in proxy.go for more flexible routing and handling of backend services.

- Update StripPrefix in ProxyConfig to "/auth".
- Add new configuration options to ProxyConfig: NotFoundRedirectURL and NoRoleRedirectURL.
- Update ProxyHandler function in proxy.go to handle not found and no role cases.
- Add error handling for backend lookup, user ID validation, and determining user roles.
- Update reverse proxy configuration to set target URL based on backend service host and path prefix.

This commit changes the routing and handling of backend services in the gateway.

* postman collection + sample fixes

* core.Start improvements

* add default password mail and update-password handler

* deploy workflow step

* some core tests

* fix hearthbeat closing prematurely

* Makefile: Add target for running database migrations

The Makefile now includes targets `m-up`, `m-down`, and `m-create` for running database migrations. The `m-up` target applies the migrations, `m-down` rolls back the migrations, and `m-create` generates a new migration file.

src/gwservices/proxy/proxy.go: Update request handling logic

The `ProxyHandler` function now uses the `Director` field instead of the `Rewrite` field in the `httputil.ReverseProxy` struct to modify the request before forwarding it. The updated logic sets the request URL, headers, and host based on the `targetURL`, `trim`, and other factors. Additionally, the `extractPathPrefix` function has been simplified to correctly extract the path prefix from the URL.

* feat: Add development requirements and deployment instructions

The commit adds development requirements for service integration and provides detailed instructions for creating the gateway service, both for free and paid services. It also includes a list of reserved routes for internal usage.

* feat: Add debug log statement in ProxyHandler

This commit adds a debug log statement in the ProxyHandler function to log the URL and host information before forwarding the request.

* extract pkg into its own lib

* feat: add ServiceAccessHandler to handle service access with authentication

This commit adds a new `ServiceAccessHandler` function to the `Proxy` struct in `proxy.go`. This handler is responsible for handling requests to services, checking for required authentication, and redirecting users if they don't have the necessary roles. The `ServiceAccessHandler` function uses the `CheckRequiredRoles` function to verify the user's roles before passing the request to the `ProxyHandler` function. The `ProxyHandler` function has been updated to receive the host and path prefix as parameters instead of extracting them from the request. The router in `router.go` has been updated to use the `ServiceAccessHandler` and the `JWTAuth` middleware for handling all paths.

* fix ablib import

* feat: update /healthcheck endpoint path

- Updated the path for the `/healthcheck` endpoint in the gateway's main file.
- The previous path was `/hc`, but it has been changed to `/healthcheck` for clarity and consistency.

refactor: Clean up unused imports and commented code

- Removed unnecessary imports and commented code in the Stripe service file.
- Removed unused imports, commented code, and unnecessary logging statements.
- These changes improve code readability and reduce clutter.

chore: Update webhook secret for test suite

- Updated the webhook secret for the test suite to a new value.
- This ensures that the tests continue to work with the new secret.
- The new secret is "test-webhook-secret".

* feat(payment): add customerCreated function

This commit adds the `customerCreated` function to handle the creation of a user when a new customer is created. It takes a JSON message as input, unmarshals it into a `stripe.Customer` struct, and then calls the `RegisterUser` function to create a new user in the system. The created user is then returned or an error is returned if any steps fail.

* Add service metadata and stripe customer id to service documentation

This commit adds the documentation for the following code changes:

- The proxy now forwards a specific field name `X-Plan-Metadata` containing the metadata defined in the bought product price. This helps the service make decisions based on the plan/product the user bought.
- The gateway also forwards the stripe customer id in `X-Stripe-Customer-Id`.
- Ensured a unique service name is used that does not contain any space or special characters.
- Updated the Dockerfile to build a standalone container, with the option to embed a database if needed.

* Add error handling for retrieving user roles

- Updated `role.go` to handle errors when retrieving user roles from the database.
- Added a check for `pgx.ErrNoRows` and return an empty `UserRole` if the user role does not exist.
- Previously, the code did not handle the case when there are no rows returned, resulting in an error.

* Add ca-certificates to Docker image and copy certificates file from builder image

This commit adds the `ca-certificates` package to the Docker image. Additionally, the certificates file `ca-certificates.crt` is copied from the `builder` image.